fix: a lot collection form basic update ok

+++ app/api/views.py
@@ -18,6 +18,7 @@ from sqlalchemy import (
     Integer,
     extract,
     or_,
+    inspect,
 )
 
 from flask.views import MethodView
@@ -33,8 +34,9 @@ from app.models import (
     Organization,
     Identification,
     MeasurementOrFact,
+    MeasurementOrFactParameter,
+    LogEntry,
     collection_named_area,
-    #CollectionNamedArea,
     get_structed_list,
 )
 from app.taxon.models import (
@@ -50,9 +52,9 @@ from .helpers_react_admin import (
     get_list_payload,
     ReactAdminProvider,
 )
-# from app.utils import (
-#    update_or_create,
-# )
+from app.utils import (
+     SAModelChange,
+ )
 
 
 @api.route('/auth', methods=['POST', 'OPTIONS'])
@@ -560,17 +562,16 @@ class UnitMethodView(MethodView):
 
     def delete(self, item_id):
         # delete a single user
-        #print (item_id, request.args)
-        obj = session.get(self.model, item_id)
-        session.delete(obj)
-        session.commit()
-        return ra_item_response(self.RESOURCE_NAME, obj)
+        if obj := session.get(self.model, item_id):
+            session.delete(obj)
+            session.commit()
+            return ra_item_response(self.RESOURCE_NAME, obj)
 
     def put(self, item_id):
         # update
-        obj = session.get(self.model, item_id)
-        obj = self._modify(obj, request.json)
-        return ra_item_response(self.RESOURCE_NAME, obj)
+        if obj := session.get(self.model, item_id):
+            obj = self._modify(obj, request.json)
+            return ra_item_response(self.RESOURCE_NAME, obj)
 
     def options(self, item_id):
         return make_cors_preflight_response()
@@ -754,9 +755,10 @@ class CollectionMethodView(MethodView):
 
     def post(self, item_id):
         # create
-        obj = Collection()
-        obj = self._update_or_create(obj, request.json)
-        return ra_item_response(self.RESOURCE_NAME, obj)
+        #obj = Collection()
+        #obj = self._update_or_create(obj, request.json)
+        #return ra_item_response(self.RESOURCE_NAME, obj)
+        return jsonify({})
 
     def delete(self, item_id):
         # delete a single user
@@ -770,28 +772,215 @@ class CollectionMethodView(MethodView):
 
     def put(self, item_id):
         # update
-        obj = session.get(self.model, item_id)
-        obj = self._update_or_create(obj, request.json)
-        return ra_item_response(self.RESOURCE_NAME, obj)
+        if obj := session.get(self.model, item_id):
+            obj = self._insert_or_update(obj, request.json)
+            return ra_item_response(self.RESOURCE_NAME, obj)
 
     def options(self, item_id):
         return make_cors_preflight_response()
 
-    def _update_or_create(self, obj, params):
+    def _insert_or_update(self, obj, payload):
         ret = {}
         # print(params, obj, obj.id, flush=True)
+        data = payload['data']
+        dirty_fields = payload['dirty']
 
+        log_entries = {}
+        collection_change = SAModelChange(obj)
+        for name, dirty in dirty_fields.items():
+            value = data[name]
+            # print('--', name, value, flush=True)
+
+            if name == 'biotopes':
+                biotope_list = []
+                log_entries['biotopes'] = {}
+                for k, v in value.items():
+                    #origin = getattr(obj, k)
+                    origin = ''
+                    for x in obj.biotope_measurement_or_facts:
+                        if x.parameter.name == k:
+                            origin = x.value
+                    log_entries['biotopes'][k] = f'{origin}=>{v}'
+                    biotope = MeasurementOrFact(collection_id=obj.id, value=v)
+                    if p := MeasurementOrFactParameter.query.filter(MeasurementOrFactParameter.name==k).first():
+                        biotope.parameter_id = p.id
+                    #if o := MeasurementOrFactParameterOption.query.filter(MeasurementOrFactParameter.name==v).first():
+                    # 不管 option 了
+                    biotope_list.append(biotope)
+
+                obj.biotope_measurement_or_facts = biotope_list
+
+            elif name == 'identifications':
+                log_entries['identifications'] = []
+                # 全部檢查 (dirtyFields 看不出是那一個 index 改的?)
+                for id_ in value:
+                    id_obj = None
+                    if iid := id_.get('id'):
+                        id_obj = session.get(Identification, iid)
+                    else:
+                        id_obj = Identification(collection_id=obj.id)
+                        session.add(id_obj)
+                        session.commit()
+
+                    mod_change= SAModelChange(id_obj)
+
+                    id_obj.date = id_['date'] if id_.get('date') else None
+                    id_obj.date_text = id_.get('date_text', '')
+                    id_obj.sequence = id_.get('sequence', '')
+                    if x := id_.get('taxon'):
+                        id_obj.taxon_id = x['id']
+                    if x := id_.get('identifier'):
+                        id_obj.identifier_id = x['id']
+
+                    changes = mod_change.check()
+                    log_entries['identifications'].append(changes)
+
+            elif name == 'units':
+                log_entries['units'] = []
+                # 全部檢查 (dirtyFields 看不出是那一個 index 改的?)
+                units = []
+                for unit in value:
+                    if unit_id := unit.get('id'):
+                        unit_obj = session.get(Unit, unit_id)
+                    else:
+                        unit_obj = Unit(collection_id=obj.id, dataset_id=1) # TODO: default HAST
+                        session.add(unit_obj)
+                        session.commit()
+
+                    unit_change = SAModelChange(unit_obj)
+                    unit_obj.accession_number = unit.get('accession_number')
+                    unit_obj.preparation_date = unit['preparation_date'] if unit.get('preparation_date') else None
+
+                    mof_list = []
+                    # 用 SAModelChange 就可以了
+                    for k, v in unit['measurement_or_facts'].items():
+                        mof = MeasurementOrFact(unit_id=obj.id, value=v)
+                        mof_change= SAModelChange(mof)
+                        if p := MeasurementOrFactParameter.query.filter(MeasurementOrFactParameter.name==k).first():
+                            mof.parameter_id = p.id
+                        #if o := MeasurementOrFactParameterOption.query.filter(MeasurementOrFactParameter.name==v).first():
+                        # 不管 option 了
+                        mof_list.append(mof)
+
+                    unit_obj.measurement_or_facts = mof_list
+                    changes = unit_change.check()
+                    log_entries['units'].append(changes)
+                    units.append(unit_obj)
+
+            else:
+                if name in ['field_number', 'collect_date', 'altitude', 'altitude2', 'longitude_decimal', 'latitude_decimal', 'longitude_text', 'latitude_text', 'locality_text', 'companion_text', 'companion_text_en']:
+                    origin = getattr(obj, name, '')
+                    log_entries[name] = f'{origin}=>{value}'
+                    setattr(obj, name, value)
+                elif name == 'collector':
+                    if value and value.get('id'):
+                        obj.collector_id = value['id']
+                    else:
+                        obj.collector_id = None
+
+                    origin = getattr(obj, name, '')
+                    #log_entries[name] = f'{origin}=>{value}'
+
+        #print('log_entries', log_entries, flush=True)
+        collection_log_entries = collection_change.check()
+        #print(collection_log_entries, log_entries, flush=True)
+        log_entries.update(collection_log_entries)
+        # print(log_entries, flush=True)
+
+        log = LogEntry(
+            model='Collection',
+            item_id=obj.id,
+            action='update',
+            changes=log_entries)
+        session.add(log)
+
+        session.commit()
+        '''
         if not obj.id:
             session.add(obj)
             session.commit()
 
-        for k, v in params.items():
-            if k == 'named_areas':
-                new_named_area_ids = [x[1] for x in v if x[1]]
+        for key, dirty in dirty_fields.items():
+            if key == 'named_areas':
+                # refresh named_data
+                new_named_area_ids = [na['id'] for _, na in data['named_areas'].items()]
                 obj.named_areas = NamedArea.query.filter(NamedArea.id.in_(new_named_area_ids)).all()
+            elif key == 'biotopes':
+                mof_list = []
+                # TOOD
+                for k, v in data['biotopes'].items():
+                    mof = MeasurementOrFact(collection_id=obj.id, value=v)
+                    mof_list.append(mof)
+                obj.biotope_measurement_or_facts = mof_list
+            elif key == 'identifications':
+                ids = []
+                for id_idx, id_value in enumerate(data['identifications']):
+                    id_obj = None
+                    if iid := id_value.get('id_identification'):
+                        id_obj = session.get(Identification, iid)
+                    else:
+                        id_obj = Identification(collection_id=obj.id)
+                        session.add(id_obj)
+
+                    if dirty[id_idx].get('identifier'):
+                        id_obj.identifier_id = id_value['identifier']['id']
+                    if dirty[id_idx].get('date'):
+                        if x := id_value.get('date'):
+                            id_obj.date = x
+                    #id_obj.date_text = id_.get('date_text')
+                    #id_obj.sequence = id_.get('sequence')
+                    #if taxon := id_.get('taxon'):
+                    #    id_obj.taxon_id = taxon.get('id')
+
+            elif key == 'units':
+                units = []
+                proxy_an_list = []
+                for unit in data['units']:
+                    if unit_id := unit.get('unit_id'):
+                        unit_obj = session.get(Unit, unit_id)
+                    else:
+                        unit_obj = Unit(collection_id=obj.id, dataset_id=1) # TODO: default HAST
+                        session.add(unit_obj)
+                        session.commit()
+
+                    if an:= unit.get('accession_number'):
+                        unit_obj.accession_number = an
+                        proxy_an_list.append(an)
+
+                    if x := unit.get('preparation_date'):
+                        unit_obj.preparation_date = x
+
+                    mofs = []
+                    for parameter_id, values in unit['measurement_or_facts']:
+                        if values and len(values) > 0:
+                            if mof_id := values[0]:
+                                mof = session.get(MeasurementOrFact, mof_id)
+                            else:
+                                mof = MeasurementOrFact(
+                                    unit_id=unit_obj.id,
+                                )
+                                session.add(mof)
+
+                            mof.option_id = values[2]
+                            mof.value_en = values[1]
+                            mof.parameter_id = parameter_id
+
+                            if mof.value_en != '':
+                                mofs.append(mof)
+
+                    unit_obj.measurement_or_facts = mofs
+                    units.append(unit_obj)
+                obj.proxy_accession_numbers = '|'.join(proxy_an_list)
+                obj.units = units
+
+            elif key in ['field_number', 'collect_date', 'companion', 'companion_en', 'locality', 'altitude', 'altitude2', 'longitude_decimal', 'latitude_decimal', 'longitude_verbatim', 'latitude_verbatim']:
+                setattr(obj, key, data[key])
+        for k, v in params.items():
             elif k == 'biotopes':
                 biotopes = []
-                for parameter_id, values in v:
+                obj.biotopes = []
+                #for x in v:
+                #    print(x, flush=True)
                     mof = None
                     if values and len(values) > 0:
                         if mof_id := values[0]:
@@ -808,7 +997,6 @@ class CollectionMethodView(MethodView):
 
                         if mof.value_en != '':
                             biotopes.append(mof)
-
                 # update by relations
                 obj.biotope_measurement_or_facts = biotopes
 
@@ -871,10 +1059,10 @@ class CollectionMethodView(MethodView):
                     if ymd != v:
                         obj.collect_date = v
             else:
-                # print(k, v, flush=True)
-                setattr(obj, k, v)
+                print('--', k, v, flush=True)
+                #setattr(obj, k, v)
+        '''
 
-        session.commit()
         '''
         # Identifications
         id_map = {x['id']: x for x in data['identifications'] if x.get('id')}
